+++ CHANGELOG.md
@@ -13,8 +13,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 
-- A couple of IntelliJ Platform errors/warnings
+- Fixed several UI/UX issues related to code completions for IDE versions starting from 233 
 - Error when adding a single file to the context
+- A couple of IntelliJ Platform errors/warnings
 
 ### Removed